MVP-6030: Update `notifi-react` to support `TenantConfigV2` (#784)
* MVP-6030: Update notifi-react to support TenantConfigV2

* MVP-6030-1: Remove redundant poc data
