+++ packages/notifi-frontend-client/lib/client/NotifiFrontendClient.ts
@@ -1094,56 +1094,7 @@ export class NotifiFrontendClient {
 
   async fetchTenantConfig(
     variables: FindSubscriptionCardParams,
-  ): Promise<TenantConfig> {
-    const query = await this._service.findTenantConfig({
-      input: {
-        ...variables,
-        tenant: this._configuration.tenantId,
-      },
-    });
-    const result = query.findTenantConfig;
-    if (result === undefined || !result.dataJson || !result.fusionEvents) {
-      throw new Error('Failed to find tenant config');
-    }
-
-    const tenantConfigJsonString = result.dataJson;
-    if (tenantConfigJsonString === undefined) {
-      throw new Error('Invalid config data');
-    }
-
-    const cardConfig = JSON.parse(tenantConfigJsonString) as CardConfigItemV1;
-    const fusionEventDescriptors = result.fusionEvents;
-
-    if (!cardConfig || cardConfig.version !== 'v1' || !fusionEventDescriptors)
-      throw new Error('Unsupported config format');
-
-    const fusionEventDescriptorMap = new Map<
-      string,
-      Types.FusionEventDescriptor
-    >(fusionEventDescriptors.map((item) => [item?.name ?? '', item ?? {}]));
-
-    fusionEventDescriptorMap.delete('');
-
-    const fusionEventTopics: FusionEventTopic[] = cardConfig.eventTypes
-      .map((eventType) => {
-        if (eventType.type === 'fusion') {
-          const fusionEventDescriptor = fusionEventDescriptorMap.get(
-            eventType.name,
-          );
-          return {
-            uiConfig: eventType,
-            fusionEventDescriptor,
-          };
-        }
-      })
-      .filter((item): item is FusionEventTopic => !!item);
-
-    return { cardConfig, fusionEventTopics };
-  }
-
-  async fetchTenantConfigV2(
-    variables: FindSubscriptionCardParams,
-  ): Promise<TenantConfigV2> {
+  ): Promise<TenantConfig | TenantConfigV2> {
     const query = await this._service.findTenantConfig({
       input: {
         ...variables,
@@ -1160,12 +1111,12 @@ export class NotifiFrontendClient {
       throw new Error('Invalid config data');
     }
     // TODO: use validator method
-    const cardConfig = JSON.parse(
-      tenantConfigJsonString,
-    ) as TenantConfigMetadata;
+    const cardConfig = JSON.parse(tenantConfigJsonString) as
+      | TenantConfigMetadata
+      | CardConfigItemV1;
     const fusionEventDescriptors = result.fusionEvents;
 
-    if (!cardConfig || cardConfig.version !== 'v2' || !fusionEventDescriptors)
+    if (!cardConfig || !fusionEventDescriptors)
       throw new Error('Unsupported config format');
 
     const fusionEventDescriptorMap = new Map<
@@ -1175,7 +1126,25 @@ export class NotifiFrontendClient {
 
     fusionEventDescriptorMap.delete('');
 
-    const fusionEventTopics: TopicMetadata[] = cardConfig.eventTypes
+    if (cardConfig.version === 'v1') {
+      const fusionEventTopics: FusionEventTopic[] = cardConfig.eventTypes
+        .map((eventType) => {
+          if (eventType.type === 'fusion') {
+            const fusionEventDescriptor = fusionEventDescriptorMap.get(
+              eventType.name,
+            );
+            return {
+              uiConfig: eventType,
+              fusionEventDescriptor,
+            };
+          }
+        })
+        .filter((item): item is FusionEventTopic => !!item);
+
+      return { cardConfig, fusionEventTopics };
+    }
+    // V2
+    const fusionEventTopics = cardConfig.eventTypes
       .map((eventType) => {
         if (eventType.type === 'fusion') {
           const fusionEventDescriptor = fusionEventDescriptorMap.get(
