--- packages/notifi-frontend-client/lib/models/SubscriptionCardConfig.ts
@@ -1,318 +1,0 @@
-/**
- * TODO: Reform tenant config
- * - Deprecated unused types
- * - increment version (TBD) - (ex. CardConfigItemV2)
- *  */
-import { Types as Gql, Types } from '@notifi-network/notifi-graphql';
-
-import { CardConfigType } from '../client';
-import { AlertFrequency, FilterOptions } from './FilterOptions';
-
-export type ValueOrRef<ValueType> =
-  | Readonly<{
-      type: 'ref';
-      ref: string | null;
-    }>
-  | Readonly<{
-      type: 'value';
-      value: ValueType;
-    }>;
-
-export type DirectPushEventTypeItem = Readonly<{
-  type: 'directPush';
-  name: string;
-  directPushId: ValueOrRef<string>;
-  tooltipContent?: string;
-  optOutAtSignup?: boolean;
-}>;
-
-export type FusionTypeBase = {
-  name: string;
-  type: 'fusion' | 'fusionToggle'; // fusionToggle is deprecated (use fusion with selectedUIType: 'TOGGLE' instead)
-  topicGroupName?: string;
-  index?: number;
-  fusionEventId: ValueOrRef<string>;
-  sourceAddress: ValueOrRef<string>;
-  tooltipContent?: string;
-  maintainSourceGroup?: boolean;
-  alertFrequency?: AlertFrequency;
-  optOutAtSignup?: boolean;
-  displayNameOverride?: string;
-};
-
-export type FusionToggleEventTypeItem = FusionTypeBase &
-  Readonly<{
-    selectedUIType: 'TOGGLE';
-  }>;
-
-export type FusionHealthCheckEventTypeItem = FusionTypeBase &
-  Readonly<{
-    selectedUIType: 'HEALTH_CHECK';
-    healthCheckSubtitle: string;
-    numberType: NumberTypeSelect;
-    checkRatios: CheckRatio[];
-    validInputRange?: ValidInputRange;
-  }>;
-
-export type ValidInputRange = {
-  max: number;
-  min: number;
-};
-
-export type FusionMultiThreshholdEventTypeItem = FusionTypeBase &
-  Readonly<{
-    selectedUIType: 'MULTI_THRESHOLD';
-    numberType: NumberTypeSelect;
-    subtitle?: string;
-    addThreshholdTitle?: string;
-  }>;
-
-export type FusionEventTypeItem =
-  | FusionToggleEventTypeItem
-  | FusionHealthCheckEventTypeItem
-  | FusionMultiThreshholdEventTypeItem;
-
-export type BroadcastEventTypeItem = Readonly<{
-  type: 'broadcast';
-  name: string;
-  broadcastId: ValueOrRef<string>;
-  tooltipContent?: string;
-  optOutAtSignup?: boolean;
-  displayNameOverride?: string;
-}>;
-
-export type HealthCheckEventTypeItem = Readonly<{
-  type: 'healthCheck';
-  name: string;
-  checkRatios: ValueOrRef<CheckRatio[]>;
-  alertFrequency: AlertFrequency;
-  tooltipContent?: string;
-  optOutAtSignup?: boolean;
-}>;
-
-export type LabelEventTypeItem = Readonly<{
-  type: 'label';
-  name: string;
-  tooltipContent?: string;
-  optOutAtSignup?: boolean;
-}>;
-
-export type TradingPairEventTypeItem = Readonly<{
-  type: 'tradingPair';
-  name: string;
-  tooltipContent?: string;
-  tradingPairs: ValueOrRef<ReadonlyArray<string>>;
-  optOutAtSignup?: boolean;
-}>;
-
-export type PriceChangeDataSource = 'coingecko';
-
-export type PriceChangeEventTypeItem = Readonly<{
-  type: 'priceChange';
-  name: string;
-  tokenIds: ReadonlyArray<string>;
-  dataSource: PriceChangeDataSource;
-  tooltipContent: string;
-  optOutAtSignup?: boolean;
-}>;
-
-export type WalletBalanceEventTypeItem = Readonly<{
-  type: 'walletBalance';
-  name: string;
-  tooltipContent?: string;
-  optOutAtSignup?: boolean;
-}>;
-
-export type CustomTypeBase = {
-  type: 'custom';
-  name: string;
-  tooltipContent: string;
-  sourceType: Gql.SourceType;
-  filterType: string;
-  sourceAddress: ValueOrRef<string>;
-  optOutAtSignup?: boolean;
-};
-
-export type CustomToggleTypeItem = Readonly<{
-  filterOptions: FilterOptions;
-  selectedUIType: 'TOGGLE';
-  optOutAtSignup?: boolean;
-}>;
-
-export type NumberTypeSelect = 'percentage' | 'integer' | 'price';
-
-export type CustomHealthCheckItem = Readonly<{
-  selectedUIType: 'HEALTH_CHECK';
-  healthCheckSubtitle: string;
-  numberType: NumberTypeSelect;
-  alertFrequency: AlertFrequency;
-  checkRatios: CheckRatio[];
-  optOutAtSignup?: boolean;
-}>;
-
-type RatiosBelow = Readonly<{
-  type: 'below';
-  ratio: number;
-}>;
-
-type RatiosAbove = Readonly<{
-  type: 'above';
-  ratio: number;
-}>;
-
-export type CheckRatio = RatiosBelow | RatiosAbove;
-
-export type CustomTopicTypeItem = CustomTypeBase &
-  (CustomToggleTypeItem | CustomHealthCheckItem);
-
-export type XMTPTopicTypeItem = {
-  type: 'XMTP';
-  name: string;
-  tooltipContent: string;
-  sourceType?: Gql.SourceType;
-  filterType: string;
-  XMTPTopics: ValueOrRef<ReadonlyArray<string>>;
-  optOutAtSignup?: boolean;
-};
-
-export type EventTypeItem =
-  | DirectPushEventTypeItem
-  | BroadcastEventTypeItem
-  | HealthCheckEventTypeItem
-  | TradingPairEventTypeItem
-  | LabelEventTypeItem
-  | PriceChangeEventTypeItem
-  | CustomTopicTypeItem
-  | FusionEventTypeItem
-  | WalletBalanceEventTypeItem
-  | XMTPTopicTypeItem;
-
-export type EventTypeConfig = ReadonlyArray<EventTypeItem>;
-export type InputType =
-  | 'WebhookUrlInput'
-  | 'WebhookHeadersInput'
-  | 'DirectPushIdInput'
-  | 'BroadcastIdInput';
-
-export type InputItem = Readonly<{
-  name: string;
-  type: InputType;
-}>;
-
-export type InputsConfig = ReadonlyArray<InputItem>;
-
-// export type ContactInfo = Readonly<{
-//   active: boolean;
-// }>;
-// export type EmailContactInfo = ContactInfo;
-// export type DiscordContactInfo = ContactInfo;
-// export type SlackContactInfo = ContactInfo;
-// export type WalletContactInfo = ContactInfo;
-export type ContactInfo = Readonly<{
-  active: boolean;
-}>;
-
-export type CountryCode = string;
-
-export type SmsContactInfo = ContactInfo &
-  Readonly<{
-    supportedCountryCodes: ReadonlyArray<CountryCode>;
-  }>;
-
-export type TelegramContactInfo = ContactInfo;
-
-export type WebhookHeaders = Readonly<Record<string, string>>;
-
-export type WebhookContactInfo = ContactInfo &
-  Readonly<{
-    url: ValueOrRef<string>;
-    headers: ValueOrRef<WebhookHeaders>;
-  }>;
-
-// export type ContactInfoConfig = Readonly<{
-//   email: EmailContactInfo;
-//   sms: SmsContactInfo;
-//   telegram: TelegramContactInfo;
-//   webhook: WebhookContactInfo;
-//   discord: DiscordContactInfo;
-//   slack: SlackContactInfo;
-//   wallet: WalletContactInfo;
-// }>;
-export type Target = // TODO: Remove type form notifi-react context. Use this instead
-  'email' | 'sms' | 'telegram' | 'discord' | 'slack' | 'wallet';
-export type ContactInfoConfig = Record<Target, ContactInfo>;
-
-export type CardConfigItemV1 = Readonly<{
-  version: 'v1';
-  id: string | null;
-  name: string;
-  eventTypes: EventTypeConfig;
-  inputs: InputsConfig;
-  contactInfo: ContactInfoConfig;
-  isContactInfoRequired?: boolean;
-  titles?: TitleSubtitleConfig;
-}>;
-
-export type TitleSubtitleConfigInactive = Readonly<{ active: false }>;
-
-export type TitleSubtitleConfigActive = Readonly<{
-  active: true;
-  editView: string;
-  previewView: string;
-  historyView: string;
-  signupView: string;
-  expiredView: string;
-  alertDetailsView: string;
-  verifyWalletsView: string;
-}>;
-
-export type TitleSubtitleConfig =
-  | TitleSubtitleConfigActive
-  | TitleSubtitleConfigInactive;
-
-export type FusionEventTopic = {
-  // The following from original fusionEventTypeItem (Legacy)
-  uiConfig: FusionEventTypeItem;
-  // The following is the respective fusionEventDescriptor
-  fusionEventDescriptor: Types.FusionEventDescriptor;
-};
-
-export type TenantConfig = {
-  cardConfig: CardConfigType; // Legacy
-  fusionEventTopics: ReadonlyArray<FusionEventTopic>;
-};
-
-export type TenantConfigV2 = {
-  cardConfig: TenantConfigMetadata;
-  fusionEventTopics: Array<TopicMetadata>;
-};
-
-// export const isTenantConfigV2 = (
-//   tenantInfo: ,
-// ): tenantInfo is TenantConfigV2 =>
-//   (tenantInfo as TenantConfigV2).cardConfig.version === 'v2';
-
-export type TopicMetadata = {
-  uiConfig: TopicUiConfig;
-  fusionEventDescriptor: Types.FusionEventDescriptor;
-};
-
-export type TopicUiConfig = {
-  name: string;
-  type: 'fusion';
-  topicGroupName?: string;
-  index?: number;
-  fusionEventId: string; // TODO: need further refactor (changed from ValueOrRef<string> to string)
-  tooltipContent?: string;
-  optOutAtSignup?: boolean;
-  displayNameOverride?: string;
-};
-
-/** v2 of CardConfigItemV1 (rename) */
-export type TenantConfigMetadata = {
-  version: 'v2'; // TODO: TBD
-  id: string;
-  contactInfo: ContactInfoConfig;
-  isContactInfoRequired?: boolean;
-  eventTypes: Array<TopicUiConfig>;
-};
