+++ packages/notifi-frontend-client/lib/models/index.ts
@@ -1,5 +1,5 @@
 export * from './FilterOptions';
 export * from './FusionEvent';
-export * from './SubscriptionCardConfig';
 export * from './SmartLinkConfig';
+export * from './TenantConfig';
 export * from './Blockchain';
